implement connect(),connectloop(),prepare(),execute()
All this so we can reconnect if a connection failure exists.
If the DB server is down, 10 retries later, the app will gracefully exit now.
If a query fails and the connection to the DB server is the issue, it will
re-try to connect and gracefully fail or re-try.
The one caveat is that if a ->prepare succeeds and then we loose the connection
a ->execute will exit(1) for now.  This can be remiedied in the future with more code, but for now this is much better than where we were before.

All of this must go in conjunction with a FDC-Db update as well.